Support multiple processor architectures
ORGANIZATION provides a top-level name for the Vagrant box hierarchy created using
this make file.

BOX_NAME now specifies the Vagrant box name, less the provider name. The provider
name is still fixed to "virtualbox" as this is the only provider used by these
Vagrant boxes.

PACKER_TEMPLATE and PACKER_VARS are both provided on the command-line, instead of
defaulting to a hard coded value in the make file.